server/all_requirements.txt
absl-py==2.1.0
astunparse==1.6.3
bcrypt==4.1.2
blinker==1.7.0
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.2
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
flatbuffers==23.5.26
gast==0.4.0
google-auth==2.28.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.60.1
h5py==3.10.0
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
jax==0.4.24
Jinja2==3.1.3
joblib==1.3.2
keras==2.12.0
libclang==16.0.6
Markdown==3.5.2
MarkupSafe==2.1.5
ml-dtypes==0.3.2
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.9.0.80
opt-einsum==3.3.0
packaging==23.2
pillow==10.2.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyJWT==2.8.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.4.1.post1
scipy==1.12.0
six==1.16.0
SQLAlchemy==2.0.27
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.3.0
typing_extensions==4.9.0
urllib3==2.2.1
Werkzeug==3.0.1
wrapt==1.14.1
zipp==3.17.0
